requirements.txt
Django>=4.2
Pillow
python-dotenv
djangorestframework
